Prices corrections

+++ controllers/packages_ajax.py
@@ -63,7 +63,7 @@ def create_price():
         return ''
     if v.price:
         try:
-            if int(v.price) > 0:
+            if float(v.price) > 0:
                 data['price'] = v.price
         except:
             return ''
@@ -207,7 +207,7 @@ def update_price():
         return ''
     if v.price:
         try:
-            if int(v.price) > 0:
+            if float(v.price) > 0:
                 data['price'] = v.price
         except:
             return ''
