+++ controllers/services_ajax.py
@@ -57,7 +57,7 @@ def create_price():
         return ''
     if v.price:
         try:
-            if int(v.price) > 0:
+            if float(v.price) > 0:
                 data['price'] = v.price
         except:
             return ''
@@ -211,7 +211,7 @@ def update_price():
         return ''
     if v.price:
         try:
-            if int(v.price) > 0:
+            if float(v.price) > 0:
                 data['price'] = v.price
         except:
             return ''
